+++ addon/app/util.js
@@ -63,6 +63,7 @@ const DEVICE_ICONS = {
   dockstream2_cordless: "💧",
   dockstream_rfid: "💧",
   one_rfid: "🐾",
+  polar: "🐾",
 };
 
 const DEVICE_REQUIRES_PET = {